# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: SIGP (Sistema Integral de Gestión de Proyectos) API
  version: 1.0.0
  description: |
    API REST completa para el Sistema Integral de Gestión de Proyectos (SIGP).

    El sistema gestiona:
    - Autenticación y autorización basada en roles
    - Planificación estratégica (PGD, OEI, OGD, OEGD, Acciones Estratégicas)
    - Proyectos y subproyectos (POI)
    - Gestión ágil (Épicas, Sprints, Historias de Usuario, Tareas)
    - Recursos humanos (Personal, Divisiones, Habilidades)
    - Notificaciones en tiempo real
    - Gestión de archivos con MinIO
    - Dashboards y métricas

  contact:
    name: API Support
    email: api@inei.gob.pe

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3010/api/v1
    description: Development server
  - url: https://api.sigp.inei.gob.pe/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido en el login
  schemas:
...
    RequestUploadUrlDto:
      type: object
      required:
        - entidadTipo
        - entidadId
        - categoria
        - nombreArchivo
        - mimeType
        - tamano
      properties:
        entidadTipo:
          type: string
          enum:
            - PROYECTO
            - SUBPROYECTO
            - ACTIVIDAD
            - ACTA_CONSTITUCION
            - ACTA_REUNION
            - DOCUMENTO
            - REQUERIMIENTO
            - CRONOGRAMA
            - INFORME_SPRINT
            - INFORME_ACTIVIDAD
            - HISTORIA_USUARIO
            - TAREA
            - SUBTAREA
            - DAILY_MEETING
            - USUARIO
          example: PROYECTO
        entidadId:
          type: integer
          minimum: 1
          example: 1
        categoria:
          type: string
          enum:
            - documento
            - evidencia
            - acta
            - informe
            - cronograma
            - avatar
            - adjunto
            - backup
          example: documento
        nombreArchivo:
          type: string
          maxLength: 255
          example: arquitectura_sistema.pdf
        mimeType:
          type: string
          maxLength: 100
          example: application/pdf
        tamano:
          type: integer
          minimum: 1
          maximum: 52428800
          description: Tamaño en bytes (máximo 50MB)
          example: 2048576
        metadata:
          type: object
          nullable: true
          example:
            fase: Diseño
            version: "1.0"
        esObligatorio:
          type: boolean
          default: false
    UploadUrlResponseDto:
      type: object
      properties:
        uploadUrl:
          type: string
          description: URL presignada para subir el archivo
          example: https://minio.inei.gob.pe/sigp/proyecto_1/arquitectura.pdf?X-Amz-Algorithm=...
        archivoId:
          type: string
          format: uuid
          description: ID del archivo creado
        objectKey:
          type: string
          description: Clave del objeto en MinIO
        bucket:
          type: string
          description: Bucket de destino
        expiresIn:
          type: integer
          description: Segundos hasta que expire la URL (usualmente 3600)
        requiredHeaders:
          type: object
          description: Headers requeridos para el PUT
          properties:
            Content-Type:
              type: string
    ConfirmUploadDto:
      type: object
      required:
        - archivoId
      properties:
        archivoId:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        checksumMd5:
          type: string
          maxLength: 32
          nullable: true
          description: MD5 checksum calculado por el cliente
    ArchivoResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombreOriginal:
          type: string
        nombreAlmacenado:
          type: string
        mimeType:
          type: string
        tamanoBytes:
          type: integer
        tamanoLegible:
          type: string
          example: 2.0 MB
        categoria:
          type: string
        estado:
          type: string
        entidadTipo:
          type: string
        entidadId:
          type: integer
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer

    # ==================== NOTIFICATION SCHEMAS ====================
...
security:
  - bearerAuth: []
paths:
...
  /upload/request-url:
    post:
      tags:
        - Storage
      summary: Solicitar URL presignada para subida de archivo
      operationId: requestUploadUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUploadUrlDto'
      responses:
        201:
          description: URL presignada generada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponseDto'
        400:
          description: Formato o tamaño no permitido
  /upload/confirm:
    post:
      tags:
        - Storage
      summary: Confirmar subida completada
      operationId: confirmUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadDto'
      responses:
        200:
          description: Subida confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchivoResponseDto'
  /upload/direct:
    post:
      tags:
        - Storage
      summary: Subida directa de archivo
      operationId: uploadDirect
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - entidadTipo
                - entidadId
                - categoria
              properties:
                file:
                  type: string
                  format: binary
                entidadTipo:
                  type: string
                  enum:
                    - PROYECTO
                    - SUBPROYECTO
                    - ACTIVIDAD
                    - USUARIO
                entidadId:
                  type: integer
                categoria:
                  type: string
                  enum:
                    - documento
                    - evidencia
                    - acta
                    - informe
                    - avatar
                metadata:
                  type: object
      responses:
        201:
          description: Archivo subido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchivoResponseDto'
...